requirements.txt
streamlit
pandas
plotly
twelvedata
google-genai